#185 feat: 생활정보 필터링 기반 비슷한 여행지 조회 전략패턴 구현
